back/requirements.txt
Django==2.0.5
django-cors-headers==2.2.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
Pillow==5.1.0
PyJWT==1.6.1
pytz==2018.4
